refactor(desktop): extract deferred launch helpers into updater/deferred-launch
Mechanical move out of updater.ts: the embedded mac shell and Windows
PowerShell wait-for-pid installer-opening scripts, their writers and
spawn plumbing, and the payload after-quit launch move to
updater/deferred-launch.ts. updater.ts re-exports the three deferred
launch input/result types; orphaned imports pruned. Region-based cut
(the PS1 heredocs contain column-zero braces, so this slice moves as one
contiguous span).
